Added in list directories and files story
There is now a 'LIST' command that the user can run.
When it is run the server will get the files/directories of the
local/server and then return it the client. The client will then print
out what has been sent to the user.
I simply printed each name out, so if it has an extension then it is a
file but if it is just a name then it is a directory